REFACTOR - Schema Type
Refactoring the schema type to not require a map iterator type to be
exposed. Far too much of the internals of the implementation were
leaking out through the abstraction.

Refactor isn't completely done, this is just a stepping stone toward
a bigger piece of work, but it's working now so commit and move on.