Nested Serialise working
That is a composite type containing a composite type produces a blob
that can be dumped using the java blob-inspector. Baby steps.

I'm fairly certain nesting the same type will break things horrible.

Also, there is very little point to the SerializerFactory at the moment
as everything is done on the ModifiableAMQPBlobImpl. This is really
messy and I don't like it but we're still playing with how
this should work so for now it's fine.

I like the idea of the factory to allow some plugability, but every
time I start thinking about what would you plug in... I run
into difficult questions about some underlaying aspects, like
the use of QPID/Proton.