More work on getting the ability to read a blob into a class (#12)
Required some refactoring of where code lived. Originally it was split across the blob and the factory. I remember that smelt odd at the time but I plowed on with it for reasons... which was clearly ya mistake as you just end up in circle dependency hell.

At this point, we can now read a nested composite